codekitchen/theapp/requirements.txt
pandas >= 0
streamlit_embedcode >= 0
textblob >= 0
